move uri.js into util subfolder, add some docs
