+++ lib/link.class.js
@@ -1,11 +1,8 @@
-/* This class represents a telehash link, containing one or more pipes (path wrappers)
- * https://github.com/telehash/telehash.org/blob/master/v3/link.md
- * @author Ryan Bennett, refactoring from mesh.js by Jeremie Miller
- */
+
 
 var hashname = require('hashname');
 var base32 = hashname.base32;
-var urilib = require('./uri');
+var urilib = require('./util/uri');
 var events = require('events');
 var util   = require('util');
 var lob    = require('lob-enc');
@@ -26,7 +23,15 @@ var log = {
  module.exports = TLink;
 
 
-
+ /**
+  * @class TLink - Telehash Link, a container for one or more @pipes (path wrappers). Most link methods are added by
+  * either bundled or third party extensions. This class mostly provides the interface by which extension functionality is
+  * added to a link (either directly or by adding an extension to the parent mesh)
+  * https://github.com/telehash/telehash.org/blob/master/v3/link.md
+  * @constructor
+  * @param {meshArgs} args - a hash of options for mesh initialization
+  * @param {function} callback - receives the link as the second argument or any error as the first
+  */
  function TLink (mesh, args, log){
 
    if (!(this instanceof TLink))
@@ -197,29 +202,37 @@ TLink.prototype.setStatus = function TLink_setStatus(err){
   this.emit('status', this.down, this)
 }
 
+
+/** set a listener for the link 'status' event, which is fired any time
+ * the link status changes.
+ * @param {function=} callback - called with 'init', 'down', 'up' as first argument, and the link as the second
+ * @return {string} 'init', 'down', or 'up'
+ */
 TLink.prototype.status = function TLink_addStatusListener(cb){
-  if(typeof cb != 'function')
-    return this.down;
+  if(typeof cb === 'function')
+  {
+    var listeners = this.listeners('status');
 
-  var listeners = this.listeners('status');
-
-  for (var i in listeners)
-    if (listeners[i] === cb)
-      return;
+    for (var i in listeners)
+      if (listeners[i] === cb)
+        return;
 
 
-  this.on('status', cb);
+    this.on('status', cb);
 
-  if(this.down != 'init'){
-    function make_status_listener(link, cb){
-      return function status_listener(){
-        link.emit('status', link.down, link);
-      };
+    if(this.down != 'init'){
+      function make_status_listener(link, cb){
+        return function status_listener(){
+          link.emit('status', link.down, link);
+        };
+      }
+      process.nextTick(make_status_listener(this, cb));
     }
-    process.nextTick(make_status_listener(this, cb));
   }
-
+  //TODO: dynamically check status when this function is called?
+  return this.down;
 }
+
 
 TLink.prototype.addPipe = function TLink_addPipe(pipe, see)
 {
@@ -294,54 +307,54 @@ TLink.prototype.router = function(router)
   return true;
 }
 
- TLink.prototype.setInfo = function TLink_setInfo(args){
-   var mesh = this._mesh;
-   // update/set json info
-   this.json = loadMeshJSON(mesh,mesh.hashname, args);
-   this.csid = hashname.match(mesh.keys, this.json.keys);
+TLink.prototype.setInfo = function TLink_setInfo(args){
+ var mesh = this._mesh;
+ // update/set json info
+ this.json = loadMeshJSON(mesh,mesh.hashname, args);
+ this.csid = hashname.match(mesh.keys, this.json.keys);
+}
+
+TLink.prototype.createExchange = function Tlink_createExchange(){
+ //first check if we already have one
+ if (this.x)
+   return true;
+
+ // no csid === no exchange
+ if (!this.csid){
+   this.x = false;
+   return false;
  }
 
- TLink.prototype.createExchange = function Tlink_createExchange(){
-   //first check if we already have one
-   if (this.x)
-     return true;
+ var exchangeOpts = {
+   csid  : this.csid
+   , key : base32.decode(this.json.keys[this.csid])
+ };
 
-   // no csid === no exchange
-   if (!this.csid){
-     this.x = false;
-     return false;
-   }
+ this.x = this._exchange(exchangeOpts);
 
-   var exchangeOpts = {
-     csid  : this.csid
-     , key : base32.decode(this.json.keys[this.csid])
-   };
+ if (!this.x && this.csid)
+   log.warn('failed to create exchange', this.json)
+ else {
+   // add the exchange token id for routing back to this active link
+   this._mesh.index[this.x.id] = this;
 
-   this.x = this._exchange(exchangeOpts);
-
-   if (!this.x && this.csid)
-     log.warn('failed to create exchange', this.json)
-   else {
-     // add the exchange token id for routing back to this active link
-     this._mesh.index[this.x.id] = this;
-
-     var mesh = this._mesh
-       , link = this;
-     this.x.sending = function TLink_x_sending(packet, pipe)
+   var mesh = this._mesh
+     , link = this;
+   this.x.sending = function TLink_x_sending(packet, pipe)
+   {
+     if(!packet) return log.debug('sending no packet',packet);
+     if(!pipe)
      {
-       if(!packet) return log.debug('sending no packet',packet);
-       if(!pipe)
-       {
-         if(link.pipes.length == 0) return log.debug('no pipes for',link.hashname);
-         pipe = link.pipes[0];
-       }
-       log.debug(mesh.hashname.substr(0,8),'delivering',packet.length,'to',link.hashname.substr(0,8),pipe.path);
-       pipe.send(packet, link, function(err){
-         if(err) log.debug('error sending packet to pipe',link.hashname,pipe.path,err);
-       });
+       if(link.pipes.length == 0) return log.debug('no pipes for',link.hashname);
+       pipe = link.pipes[0];
      }
+     log.debug(mesh.hashname.substr(0,8),'delivering',packet.length,'to',link.hashname.substr(0,8),pipe.path);
+     pipe.send(packet, link, function(err){
+       if(err) log.debug('error sending packet to pipe',link.hashname,pipe.path,err);
+     });
    }
  }
+}
 
 TLink.prototype.initialize = function TLink_initialize(args){
   // if the link was created from a received packet, first continue it
