+++ lib/mesh.class.js
@@ -6,7 +6,7 @@ var lob = require('lob-enc');
 var stringify = require('json-stable-stringify');
 var TLink = require('./link.class');
 var Pipe = require('./pipe.class');
-var urilib = require('./uri');
+var urilib = require('./util/uri');
 var handshakelib = require("./util/handshake")
 var loadMeshJSON = require('./util/json').loadMeshJSON;
 module.exports = Mesh;
@@ -15,6 +15,7 @@ var log = {};
 
 
 /**
+ * @typedef Mesh
  * @class Mesh - Telehash Mesh Node
  * @constructor
  * @param {meshArgs} args - a hash of options for mesh initialization
@@ -469,7 +470,6 @@ Mesh.prototype.discovered = function(to)
  * if the link argument is ommited, this function sets the behaivior of the local node
  * @param {TLink=} [link=The local Mesh] - the remote link to use as a router
  * @param {Boolean=} [isRouter=true] - whether to declare the target as a router;
- *
  */
 Mesh.prototype.router = function(link, isRouter)
 {
