+++ lib/mesh.js
@@ -5,7 +5,7 @@ var base32 = hashname.base32;
 var lob = require('lob-enc');
 var stringify = require('json-stable-stringify');
 var Pipe = require('./pipe.class');
-var urilib = require('./uri');
+var urilib = require('./util/uri');
 var Mesh = require('./mesh.class');
 
 // export some of these to the whole package for apps to use directly as telehash.*
